feat: rounding speed limit value on output page and adding slope and …

+++ pages/train_speed_page.cpp
@@ -100,11 +100,12 @@ void TrainSpeedPage::updateCharts() {
 
 void TrainSpeedPage::setParameterValue() {
   m_inputWidgets["Max Speed"]->setValue(0);
-  m_inputWidgets["Max Speed"]->setValue(m_trainSimulation->findMaxSpeed());
+  m_inputWidgets["Max Speed"]->setValue(
+      static_cast<int>(m_trainSimulation->findMaxSpeed()));
 }
 
 void TrainSpeedPage::setStaticParameterValue() {
   m_inputWidgets["Static Max Speed"]->setValue(0);
   m_inputWidgets["Static Max Speed"]->setValue(
-      m_trainSimulation->findMaxSpeed());
+      static_cast<int>(m_trainSimulation->findMaxSpeed()));
 }
